added Passwordable modules to lib
